refactor(cli): move master genie detection to update.ts command
- Move 70+ lines of master genie logic from genie-cli.ts to update.ts
- Add up-to-date check before installing (prevents unnecessary reinstalls)
- Reduce genie-cli.ts from 1508 to 1439 lines (still over 1000 limit)
- Enforce new amendment: files must be < 1000 lines where possible

Fixes version paradox where running genie update twice performs install both times even when already up-to-date.
